requirements.txt
pandas
pyodbc
snowflake-connector-python
